+++ plugins/fraiseql-confiture-pggit/src/confiture_pggit/cli/__init__.py
@@ -1,0 +1,1 @@
+"""pgGit's command groups, added to ``confiture`` by :func:`confiture_pggit.register`."""